Created an Inbox subcomponent
